openlobby-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.2
httpx==0.27.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
supabase==2.6.0
elasticsearch[async]==8.15.0
beautifulsoup4==4.12.3
lxml==5.3.0
modal==0.67.0

